requirements.txt
python-dotenv==1.0.1
request==2.32.3
flask==3.0.3
flask-cors==4.0.1
langchain==0.2.7
langchain-community==0.2.7
chromadb==0.5.4